requirements.txt
asyncssh>=2.14.2
mcp>=1.3.0
pyyaml>=6.0.1
pymodbus>=3.6,<4
pyserial>=3.5
